+++ MangaMark_Vue/src/factory/websiteFactory.ts
@@ -41,6 +41,7 @@ export class WebsiteFactory {
         return new AqumangaWebsite()
       case domain.includes('toptruyen'):
       case domain.includes('doctruyen3q'):
+      case domain.includes('truyen3q'):
         return new ToptruyenWebsite()
       case domain.includes('manganato'):
         return new ManganatoWebsite()