+++ MangaMark_Vue/src/js/website/cmangaWebsite.ts
@@ -3,6 +3,8 @@ import CacheMangaApi from "../utils/cacheMangaApi";
 import delay from "../utils/delay";
 import getChapterNumber from "../utils/getChapterNumber";
 import handleChapterJump from "../utils/handleChapterJump";
+import removeCookie from "../utils/removeCookie";
+import removeElements from "../utils/removeElements";
 import toDataString from "../utils/toDataString";
 import Website from "./website";
 
@@ -53,13 +55,8 @@ export default class CmangaWebsite implements Website {
   };
 
   blockAds(): void {
-    const fbRoot = document.getElementById('fb-root')
-    if (fbRoot) {
-      fbRoot.nextElementSibling?.remove()
-    }
-    const popupContent = document.getElementById('popup_content')
-    if (popupContent) {
-      popupContent.remove()
-    }
+    removeElements('.pr_module');
+    removeElements('img', 'cmangapi');
+    removeCookie('ads_num')
   }
 }